dediprog: Split read/write command preparation by protocol
There has been a forest of `if`s growing inside prepare_rw_cmd(). And
the V3 protocol allows for more flexibility that would result in even
more branches if we'd continue adding to this function.

Instead, split it into one function per protocol version  and provide
a pointer to it in our context structure. This will allow us to adapt
the V3 function more easily and only sacrifices a little code sharing
between V2 and V3.  We also let it return the number of bytes for the
command packet and define a maximum size, to simplify the API.

Change-Id: Ibe9da3d3f1aac74309b89f840d9ce9e6e7978405
Signed-off-by: Nico Huber <nico.h@gmx.de>
Reviewed-on: https://review.sourcearcade.org/c/flashprog/+/126
Reviewed-by: Arthur Heymans <arthur@aheymans.xyz>