manibuilder: Use `test_build.sh'
Use `test_build.sh' to test builds with both Make and Meson, and to
compare the results. To succeed, the $(MESONARGS) have to mimic the
behavior of `CONFIG_EVERYTHING=yes'.

In the long run, we might want to use the version shipped with mani-
builder instead of the tested flashprog version. Otherwise, we won't
be able to test older revisions of flashprog with newer manibuilder.

Change-Id: Ife327e2f3e2fa9acbbfd1fa51ec0aa3c3f2489b8
Signed-off-by: Nico Huber <nico.h@gmx.de>
Reviewed-on: https://review.sourcearcade.org/c/flashprog/+/144