meson: Check for CPU families with known raw mem access
Same as the Makefile does with $(DEPENDS_ON_RAW_MEM_ACCESS), we should
tag all the programmers that require raw memory access. Otherwise, we
might compile these drivers for architectures that require explicit
memory barriers that we haven't implemented.

As there seems to be no notion of intersection in Meson, we have to make
`cpu_families' a list of lists and check them all.

Change-Id: I3bd6bd7b5cb871a506178496241bacdeb40febcf
Signed-off-by: Nico Huber <nico.h@gmx.de>
Reviewed-on: https://review.sourcearcade.org/c/flashprog/+/131
Reviewed-by: Arthur Heymans <arthur@aheymans.xyz>