Properly clear erase-block selection when bigger block is chosen
When we choose to erase a bigger block over many smaller sub blocks,
we should clear the `selected' status  of sub blocks recursively. If
not, we risk to miss sub blocks that haven't been consolidated in an
intermediate level of erase blocks.

For instance, considering a flash chip with three erase functions:

    32x   4KiB sectors
     4x  32KiB blocks
     1x 128KiB chip

With the following erase pattern at the lowest, 4KiB level (`x' for
a block that is selected for erase, `.' for one left untouched):

    xxx..xxx ..x..x.. xxxxxxxx x.x.....    4KiB

At the intermediate level, we would consolidate those 32KiB blocks
that get more than half of their sub blocks erased:

    ........ ..x..x.. ........ x.x.....    4KiB
        x        .        x        .      32KiB

Still more than half of the flash gets erased, so we would consoli-
date this to a full chip erase:

    ........ ........ ........ ........    4KiB
        .        .        .        .      32KiB
                     x                   128KiB

The current implementation, however, wouldn't clear the remaining
selected blocks at the 4KiB level.

Change-Id: I90181f9c53ca9aa06d353ba6dd4072efbebc260c
Signed-off-by: Nico Huber <nico.h@gmx.de>
Reviewed-on: https://review.sourcearcade.org/c/flashprog/+/161
Reviewed-by: Angel Pons <th3fanbus@gmail.com>
Reviewed-by: Arthur Heymans <arthur@aheymans.xyz>