ichspi: Split ICH7 init out
The original, ICH7 init only shared about three lines with the newer,
ever growing ICH9+ init. That's not worth an indentation level in an
endlessly long function, so split it out.

We introduce a kind of "breakpoint" into the `ich_chipset' enum:

  SPI_ENGINE_ICH9

This marks all chipset entries below it as supporting this code path
and should help to avoid long `case' lists.

Tested read/erase/write on ThinkPad T60 (ICH7).

Change-Id: I41e46d12e02c1343e636b47b2378db86e76af95e
Signed-off-by: Nico Huber <nico.h@gmx.de>
Reviewed-on: https://review.sourcearcade.org/c/flashprog/+/173
Reviewed-by: Arthur Heymans <arthur@aheymans.xyz>