ch347_spi: Add 'spimode' parameter
This allows the SPI mode (clock polarity and phase) of the CH347 to be
selected. By default mode 0 is used, as most flash chips are compatible
with this mode. I have noticed that the CH347 is able to work at higher
clock speeds with some chips when set to mode 1, despite the chip not
officially having support for this configuration.

Change-Id: I7938519e23e9e014c016f9d7f130d1ac191a09fa
Signed-off-by: Nicholas Chin <nic.c3.14@gmail.com>
Reviewed-on: https://review.sourcearcade.org/c/flashprog/+/244
Reviewed-by: Nico Huber <nico.h@gmx.de>
Tested-by: Nico Huber <nico.h@gmx.de>