chipset_enable: Factor PCH100 hidden-spidev workaround out
Intel mandates that the SPI PCI device,  usually 00:1f.5, is hidden
on all 22nm PCHs. This applies to the whole 100 and 200 series PCHs
as well as some special 300 and 400 series PCHs, and the respective
server PCHs and small-core SoCs.

To cope with the hidden PCI device, we match the LPC PCI device and
then crudely assume that the SPI device exists too (only its vendor
and device IDs are hidden). We don't need this workaround for newer
generations where the PCI device isn't hidden anymore,  hence split
it out.

Tested read on Alder Lake P, read/erase/write on CM246 (Cannon Point).

Change-Id: I77b5240b99015ecf56773f4a34436cfd3c83bdf6
Signed-off-by: Nico Huber <nico.h@gmx.de>
Reviewed-on: https://review.sourcearcade.org/c/flashprog/+/249