chipset_enable: Remove hidden-spidev workaround for TGP+
When the 500 series PCH (Tiger Point) support was added, we did
not know if all firmware will be consistent in the decision not
to hide the SPI PCI devices.  A few years later, we can confirm
that it is: Grepping[1] through the `linuxhw' database[2] shows
that we actually gain some hardware support,  when matching the
SPI PCI device directly.

For Alder Lake and Raptor Lake, we already used the PCI IDs of
the SPI device.

Tested read on Alder Lake P, read/erase/write on RM590E (Tiger
Point) which was previously not detected.

[1] diff -u <(git grep -lE '1f.5.*8086:(a0a4|43a4)' | grep -v README) \
            <(git grep -lE '1f.0.*8086:(a08[123678]|438[1-df])' | grep -v README)
[2] https://github.com/linuxhw/LsPCI/

Change-Id: I2474d94be53fe01f8bd01d924098fa28fd43d657
Signed-off-by: Nico Huber <nico.h@gmx.de>
Reviewed-on: https://review.sourcearcade.org/c/flashprog/+/251