udev rules: Use `uaccess' tag instead of `plugdev' group
Not all systems have a `plugdev' group.  The `uaccess' tag, however,
seems to be the standard these days  (seems to exist for about a de-
cade on major distros). We rename the file to have a higher priority
by default. At least on ArchLinux, it has to precede `73-seat-late'.

Tested on ArchLinux (local session) and Raspbian (remote session).

Change-Id: Iadcc74b32b92660fbf86b235e65692369535f7a6
Signed-off-by: Nico Huber <nico.h@gmx.de>
Reviewed-on: https://review.sourcearcade.org/c/flashprog/+/266