ichspi: Add Intel Arrow Lake support
ARL looks much like a desktop version of Meteor Lake. Hardware registers
seem to be the same, and the descriptor mostly differs in strap settings
(as far as we are concerned).

Odd enough, the old (pre 500 series) format for processor straps is used
again. For the descriptor detection, we shuffle the old default for Ibex
Peak around, and make Arrow Lake the default for everything with over 80
PCH traps.

Tested `ich_descriptors_tool' output for a GIGABYTE Z890M GAMING X BIOS.

Documents used:
  * Intel® Core™ Ultra 200S Series Processors Datasheet, Volumes 1 and 2
  * Arrow Lake-S and Arrow Lake-HX Client Platform
    SPI Programming Guide

Change-Id: Ibaaeb896273eed3806561ba8c01d89770d27ff18
Signed-off-by: Nico Huber <nico.h@gmx.de>
Reviewed-on: https://review.sourcearcade.org/c/flashprog/+/270