dediprog: Set i/o lines with protocol v2+
When adding support for the v2 protocol, we left a FIXME open
instead of setting the i/o lines as before.  Looking into the
`dpcmd` source code confirms  that the i/o values are now set
in the lower bits of `wValue` of the USB request.

Traces of Dediprog's Windows application with an SF600Plus-G2
show that it first sends a request with `wValue = 0x8001` and
`wIndex = 0x0021` during initialization, then later `wValue =
0x0L09` by default (L being the LED setting).

In a ticket[1], test results showed that setting 0x09 fixes
spurious high-levels on the MISO line with newer SF100 ver-
sions and certain flash chips.

Tests with an SF600Plus-G2 confirmed that the magic 0x9 sets
CS2 and IO2 high. Adapters map IO2 to IO4 of an SF100 inter-
face, which again is expected to be high.

[1] https://ticket.coreboot.org/issues/609

Change-Id: Iba5e113adb607a92802cb93a121987920f507fad
Signed-off-by: Nico Huber <nico.h@gmx.de>
Reviewed-on: https://review.sourcearcade.org/c/flashprog/+/349